move files to new location
